call[fs]cgi: At least for now, simply block when connecting to handler.
It's not optimal, but the naive mtio implementation has the thundering-herd
problem, and this also creates a natural choke-point, transmitting the request
buffering responsibility to the parent handler where it belongs.

The more ideal solution would probably be to let running requests continue
being handled while serializing receving incoming requests and connecting to
the request handler.